% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAsE2toolkit.R
\docType{package}
\name{GAsE2toolkit}
\alias{GAsE2toolkit}
\title{GAsE2toolkit}
\description{
Automated design through the method of genetic algorithms
for distribution networks taking into account cost optimization
and pressure restrictions.
}
\author{
Ayala Bizarro Ivan \email{ivan.ayala@unh.edu.pe}
}
