% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open.R
\name{Open_epanet2toolkit}
\alias{Open_epanet2toolkit}
\title{Initialize the program}
\usage{
Open_epanet2toolkit()
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
more detailed description
}
\examples{
Open_epanet2toolkit()

}
